ENH: Use pose matrices instead of camera parameters for VR controllers
It is easier to manage controller pose as a matrix instead of the (pos,wxyz,ppos,wdir) variables that the ConvertPoseToWorldCoordinates function calculates. A function called ConvertOpenVRPoseToMatrices was added that calculates the matrix. It is used in various places in addition or instead of those variables.
A PhysicalToWorldMatrixModified event is added to vtkOpenVRRenderWindow to enable notifications to the application when the physical scaling, translation, etc. changes.

Small changes:
- Fix incorrect sign of VR physical to world matrix translation
- Fix gesture end event invocation (Trigger button release was checked instead of Grip, as it is done for press)
- Add comments for physical to world ivars and controller pose camera style variables for easier understanding
- Remove unused PoseTransform member variable from vtkOpenVRRenderWindowInteractor