ENH: Add timer in vtkInteractorStyle3D to allow for smooth flying
Dolly translation speed can be specified in m/s, to achieve a constant flying speed that is not dependent on frame rate or physical scale. Default is 1.6666 m/s, which corresponds to walking speed (6 km/h). Do not dolly if left or right side of the trackpad is pressed instead of up/down.