[backport] Various fixes for VR interaction
Fix an issue with the QTWidgetWidget that could cause
parts of it to be unpainted (black) when shown

Fix an issue where probed cells were highlighted but not
taking into account the actors usermatrix

Fix an issue with the VRRenderWindow not setting the
correct buffer values.

Fix an issue with the position prop code in the 3d
interactor style that was causing multiple props
to share the same user matrix, causing them to move together
when any one of them was grebbed.